add protection endpoints for users products orders auth

+++ app/domains/orders/router.py
@@ -1,17 +1,14 @@
 # app/domains/orders/router.py
 
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from core.db.postgres import get_async_session
 
-from domains.auth.dependencies import get_current_user
+from domains.auth.dependencies import get_current_admin, get_current_user
 from domains.users.models import User
 
-from .schemas import (
-    OrderCreate,
-    OrderRead,
-)
+from .schemas import OrderCreate, OrderRead
 from .service import OrderService
 
 
@@ -42,11 +39,10 @@ async def get_my_orders(
     response_model=list[OrderRead],
 )
 async def get_all_orders(
+    admin: User = Depends(get_current_admin),
     session: AsyncSession = Depends(get_async_session),
 ):
-    return await service.get_all_orders(
-        session=session,
-    )
+    return await service.get_all_orders(session=session)
 
 
 @router.get(
@@ -55,12 +51,21 @@ async def get_all_orders(
 )
 async def get_order(
     order_id: int,
+    current_user: User = Depends(get_current_user),
     session: AsyncSession = Depends(get_async_session),
 ):
-    return await service.get_order(
+    order = await service.get_order(
         session=session,
         order_id=order_id,
     )
+
+    if order.user_id != current_user.id and not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not enough permissions",
+        )
+
+    return order
 
 
 @router.post(
@@ -88,14 +93,28 @@ async def add_item(
     order_id: int,
     product_id: int,
     quantity: int,
+    current_user: User = Depends(get_current_user),
     session: AsyncSession = Depends(get_async_session),
 ):
+    order = await service.get_order(
+        session=session,
+        order_id=order_id,
+    )
+
+
+    if order.user_id != current_user.id and not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not enough permissions",
+        )
+
     return await service.add_item_to_order(
         session=session,
         order_id=order_id,
         product_id=product_id,
         quantity=quantity,
     )
+
 
 @router.delete(
     "/{order_id}/items/{product_id}",
@@ -105,8 +124,21 @@ async def remove_item(
     order_id: int,
     product_id: int,
     quantity: int | None = None,
+    current_user: User = Depends(get_current_user),
     session: AsyncSession = Depends(get_async_session),
 ):
+    order = await service.get_order(
+        session=session,
+        order_id=order_id,
+    )
+
+
+    if order.user_id != current_user.id and not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not enough permissions",
+        )
+
     return await service.remove_item_from_order(
         session=session,
         order_id=order_id,
@@ -121,8 +153,20 @@ async def remove_item(
 )
 async def delete_order(
     order_id: int,
+    current_user: User = Depends(get_current_user),
     session: AsyncSession = Depends(get_async_session),
 ):
+    order = await service.get_order(
+        session=session,
+        order_id=order_id,
+    )
+
+    if order.user_id != current_user.id and not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not enough permissions",
+        )
+
     await service.delete_order(
         session=session,
         order_id=order_id,
